perf(main.py): parralelized processing for up to 100x improvement
this is just very io dependant (api calls and wait), so easily parallelized. then the findmultiplication file was used to check the logs for how much of the rate limit was being used and adjust the number based off of that. in the future will just use the actual openai lib to do that on the fly